chore: move stories to the right folder
