+++ app/rne-theme/text.stories.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 import { storiesOf } from "@storybook/react-native"
-import { Story, StoryScreen, UseCase } from "../views"
+import { Story, StoryScreen, UseCase } from "../../storybook/views"
 import { Text } from "@rneui/themed"
 import colors from "@app/rne-theme/colors"
